takes taks from first epic
